Disable tests on windows
